491PythonBackend-master/requirements.txt
beanie==1.22.6
certifi==2023.7.22
fastapi==0.103.1
passlib==1.7.4
pydantic==2.3.0
pydantic_settings==2.0.3
pymongo==4.5.0
python_jose==3.3.0
starlette==0.31.1
typing_extensions==4.8.0
uvicorn==0.23.2